se agrega cambio de PathVarieble

+++ ms-ne-emision-poliza-tps-pga/src/main/java/pe/com/pacifico/kuntur/model/Usuario.java
@@ -25,7 +25,7 @@ public class Usuario {
 
   @Id
   @Column(name = "ID_USUARIO")
-  private Long idUsuario;
+  private int idUsuario;
 
   @Column(name = "CORREO")
   private String correo;
